creatoe entry point modificando file index.js
